requirements.txt
Flask==2.1.2
keras==2.9.0
numpy==1.23.0
opencv_python==4.6.0.66
Pillow==9.1.1
tensorflow-cpu
gunicorn
